move logging to app level
