--- app/tools/spotify/__init__.py
@@ -1,3 +1,0 @@
-from app.tools.logger import setup_logger
-
-logger = setup_logger(__name__)
