+++ app/tools/spotify/add_tracks.py
@@ -1,7 +1,7 @@
 import json
 
 import requests
-from app.tools.spotify import logger
+from app import logger
 from app.tools.spotify.config import base_url
 from app.tools.spotify.helpers import get_headers
